Refactor Peer and PeerConfig
Renames and restructures the confusingly named classes "Peer" and
"PeerConfig" into Peer -> PeerConnection and PeerConfig -> Peer, as that is
what the classes are supposed to do. The old "PeerConfig" class represented
an actual peer, while the incorrectly named "Peer" class represented only a
connection to a peer.

Also moves everything "peer" related to peer.py, as that is where they should
be.