Change behaviour of accessing defined but unset AVPs as attributes
Previously, attempting to access a known AVP for one of the
`DefinedMessage` subclasses that was not set would raise an `AttributeError`.
Changes this behaviour, by returning `None` in such cases.

AVPs *not* defined in one of the subclasses that are also not set still
continue to raise an exception when being accessed.