Detect retransmissions and reject if already answered
Adds support for rfc6733, 5.5.4, by recording the end-to-end identifier
and origin-host of each outgoing answer message, and then checking for
the existence of an end-to-end identifier in the recorded list, when a
message with the T-flag set is received. A detected retransmission is
declined with DIAMETER_UNABLE_TO_COMPLY, though this may be wrong.

Due to the need of being able to record end-to-end identifiers of
outgoing messages, *every* message sent out by a Node is now routed
though `send_message`, instead of dumping them directly into the work
queue of each connection.

Closes #1.