Take snapshots of node.node_statistics at 60 second intervals
For #9, starts a background thread at node startup, which takes a
snapshot (dataclass-as-dict) of `node_statistics` values every 60
seconds and keeps 1440 of snapshots, i.e. 24 hours.

Changes the way the cumulated `node_statistics` is calculated to
work on deep copies of peer statistics counter attributes, as those
counters are guaranteed to change size while being iterated.