Extends AvpTime usage to year 2014
Implements a fix described in rfc2020 section 3
https://datatracker.ietf.org/doc/html/rfc2030#section-3

which extends the NTP timestamp format to be useful even after the
year 2036 integer overflow would take place. The fix picks a date
in 1968 (first second where the most significant bit in the integer
seconds part is set), and considers every date before the cutoff to
be in the range of 2036-2104, while every date after the cutoff is
in the year 1968-2036 range.

This also makes rendering past timestamps, before 1968, impossible.

Closes #8.