Fix routing between multiple realms
Removes the restriction that sets each app only for the realm that the
diameter Node has been configured with. Now sets the app/peer
combination(s) for each peer realm separately, effectively allowing the
Node to handle incoming requests from more than one realm at the same
time.

The realm used in CER remains as the value that the node is configured
with.

Relates to #10.