Detect interpreter shutdown when spawning an application thread
A race condition may cause an Application to attempt to spawn a new
thread for handling an incoming message, while the main python
interpreter is already shutting down. In Python 3.12 this now raises an
RuntimeError, while previous python versions ignored it silently.

This change discards the message and produces a warning in the logs.

Relates to #11.